Create a file to test every func
